requirements.txt
python-dotenv~=1.0.1
langchain-openai~=0.3.8
langgraph~=0.3.18
langchain-core~=0.3.47
typing_extensions~=4.12.2
langchain-community~=0.3.20
pydantic~=2.10.6